worker_tracking/requirements.txt
paho-mqtt==1.6.1
tomli==2.0.1
pyzmq==23.2.1
pyudev==0.22.0
evdev==1.3.0